python_gen/cities/poubelle/requirements.txt
airtight==0.2.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
cattrs==25.1.1
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
encoded_csv==0.2
fonttools==4.58.4
h11==0.16.0
haversine==2.9.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.8
lxml==5.4.0
matplotlib==3.10.3
mypy_extensions==1.1.0
numpy==2.3.0
ollama==0.5.1
packaging==25.0
pandas==2.3.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
-e git+https://github.com/isawnyu/pleiades.datasets@0b5acc2179dfad9815ddf6f1dd5c5d1d9ad72f00#egg=pleiades.datasets
pluggy==1.6.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pytest==8.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-cache==1.2.1
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
textnorm==1.2
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
url-normalize==2.2.1
urllib3==2.4.0
